requirements.txt
pandas
GitPython
progress
